requirements.txt
absl-py==2.1.0
astunparse==1.6.3
boto3==1.36.12
botocore==1.36.12
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
copulas==0.12.1
ctgan==0.10.2
cycler==0.12.1
deepecho==0.6.1
et-xmlfile==2.0.0
Faker==35.2.0
filelock==3.17.0
flatbuffers==25.1.24
fonttools==4.55.8
fsspec==2025.2.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.70.0
h5py==3.12.1
idna==3.10
imbalanced-learn==0.13.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
keras==2.15.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
namex==0.0.8
narwhals==1.25.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
Pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
rdt==1.13.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.11.2
scikit-learn==1.3.2
scipy==1.15.1
sdmetrics==0.18.0
sdv==1.17.4
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
sympy==1.13.1
tensorboard==2.15.0
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
torch==2.0.1
tqdm==4.67.1
typing-extensions==4.5.0
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.12.0